svc/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.6.2
psycopg2==2.7.4
gunicorn==19.8.1
flask_restplus
